requirements/local.txt
-r base.txt
psycopg2==2.9.4
flake8==5.0.4
black==22.10.0
isort==5.10.1
argh==0.26.2
PyYAML==6.0
watchdog==2.1.9
selenium==2.28.0
webtest
django_webtest